rename: 커스텀 요소들 custom- 으로 폴더 네이밍 통일
