+++ frontend/src/app/design-system-test/page.tsx
@@ -2,7 +2,7 @@
 
 import { Button, Menu, MenuContent, MenuList, MenuTrigger } from '@wanteddev/wds';
 import { IconAgent, IconAiReview, IconPlus } from '@wanteddev/wds-icon';
-import { CustomMenu } from '@/components/commons/custom-menu/CustomMemu';
+import { CustomMenuItem } from '@/components/commons/custom-menu/CustomMemuItem';
 
 const Demo = () => {
   return (
@@ -14,15 +14,15 @@ const Demo = () => {
 
         <MenuContent sx={{ width: '154px' }}>
           <MenuList>
-            <CustomMenu value="item1" icon={<IconPlus />} onClick={() => console.log('test')}>
+            <CustomMenuItem value="item1" icon={<IconPlus />} onClick={() => console.log('test')}>
               Item 1
-            </CustomMenu>
-            <CustomMenu value="item2" icon={<IconAiReview />}>
+            </CustomMenuItem>
+            <CustomMenuItem value="item2" icon={<IconAiReview />}>
               Item 2
-            </CustomMenu>
-            <CustomMenu value="item3" icon={<IconAgent />}>
+            </CustomMenuItem>
+            <CustomMenuItem value="item3" icon={<IconAgent />}>
               Item 3
-            </CustomMenu>
+            </CustomMenuItem>
           </MenuList>
         </MenuContent>
       </Menu>
